Fix regression in simulator.
When I refactored the simming loop in Simulator::simulateOnePosition() using
a ranged for, I mis-translated the value of the playerNumber variable.
This resulted in the results of some plays being double-weighted.

This fixes a bug where a 2-ply sim on the opening rack ACINPU? will favor
non-bingo plays like UP or CUP rather than bingo plays like PANICUm or
PUrANIC.  The Championship Player uses a 2-ply sim, so the Championship Player
could make some bone-headed plays with this regression.