Fix devicePixelRatios on tile pixmap after dragging it.
Apparently, when you serialize a pixmap through a QDataStream(), it loses
track of its devicePixelRatio setting.  End rule was that if, on a HiDpi
screen, you dragged and dropped a tile within the rack, you'd get a tile that
was larger than it should have been by a factor of the default pixel ratio.

Solution is to simply reapply the pixel ratio to the newly created Pixmap
after deserializing it in the drag-drop code.